Port standalone fixes from upstream (kdomanski/iso9660)
The two forks diverged years ago and a direct merge is not viable (both
rewrote the writer/reader and added Rock Ridge independently; our RR/SUSP
code is already a superset of upstream's). These are the standalone,
still-applicable fixes ported into the fork's design:

- DirectoryEntry.ExtentLength is now uint32 instead of int32, decoded via
  UnmarshalUint32LSBMSB, so extents of 2 GiB and above no longer overflow
  to a negative size. (upstream 0ab0fa9, aarzilli)
- Decoding a UDF (BEA01) volume now returns a specific ErrUDFNotSupported
  error instead of a generic identifier-mismatch message. (upstream
  7d7a089, kdomanski)
- AddLocalFile/AddLocalDirectory no longer silently follow symlinks: a
  symlink now returns a clear error via os.Lstat. (upstream f09543e,
  kdomanski)
- AddLocalDirectory normalizes path separators with filepath.ToSlash so
  in-image paths are correct on Windows. (adapted from upstream 48115dd)

Added tests for the symlink rejection and the UDF error path.

+++ image_reader.go
@@ -182,7 +182,7 @@ func (f *File) GetAllChildren() ([]*File, error) {
 	baseOffset := uint32(f.de.ExtentLocation) * sectorSize
 
 	buffer := make([]byte, sectorSize)
-	for bytesProcessed := uint32(0); bytesProcessed < uint32(f.de.ExtentLength); bytesProcessed += sectorSize {
+	for bytesProcessed := uint32(0); bytesProcessed < f.de.ExtentLength; bytesProcessed += sectorSize {
 		if _, err := f.ra.ReadAt(buffer, int64(baseOffset+bytesProcessed)); err != nil {
 			return nil, nil
 		}
